+++ Sources/VComponents/Components/Containers/Disclosure Group/VDisclosureGroupLabel.swift
@@ -8,4 +8,5 @@
 import SwiftUI
 
 // MARK: - V Disclosure Group Header Label
+@available(iOS 14.0, *)
 typealias VDisclosureGroupLabel = GenericLabel_TitleCustom