+++ Sources/VComponents/Components/Containers/Disclosure Group/VDisclosureGroupState.swift
@@ -10,7 +10,9 @@ import VCore
 
 // MARK: - V Disclosure Group State
 /// Enum that represents state, such as `collapsed` or `standard`.
+@available(iOS 14.0, *)
 public typealias VDisclosureGroupState = GenericState_CollapsedExpanded
 
 // MARK: - V Disclosure Group Internal State
+@available(iOS 14.0, *)
 typealias VDisclosureGroupInternalState = GenericState_CollapsedExpandedDisabled