+++ Sources/VComponents/Components/Messages/Toast/VToastExtension.swift
@@ -9,6 +9,7 @@ import SwiftUI
 import VCore
 
 // MARK: - Bool
+@available(iOS 14.0, *)
 extension View {
     /// Presents toast when `Bool` is `true`.
     ///
@@ -58,6 +59,7 @@ extension View {
 }
 
 // MARK: - Item
+@available(iOS 14.0, *)
 extension View {
     /// Presents toast using the item as data source for content.
     ///
@@ -121,6 +123,7 @@ extension View {
 }
 
 // MARK: - Presenting Data
+@available(iOS 14.0, *)
 extension View {
     /// Presents toast when `Bool` is `true` using data to produce content.
     ///
@@ -189,6 +192,7 @@ extension View {
 }
 
 // MARK: - Error
+@available(iOS 14.0, *)
 extension View {
     /// Presents toast when `Bool` is `true` using `Error`
     ///