+++ Sources/VComponents/Helpers/Architectural Pattern Helpers/Alert/VAlertExtension_Parameters.swift
@@ -8,6 +8,7 @@
 import SwiftUI
 
 // MARK: - V Alert Extension (Parameters)
+@available(iOS 14.0, *)
 extension View {
     /// Presents `VAlert` when `VAlertParameters` is non-`nil`.
     ///
@@ -15,11 +16,12 @@ extension View {
     /// For additional info, refer to [documentation](https://github.com/VakhoKontridze/VCore/blob/main/Sources/VCore/Helpers/Architectural%20Pattern%20Helpers/SwiftUI/Alert/AlertExtension.swift) .
     public func vAlert(
         id: String,
+        uiModel: VAlertUIModel = .init(),
         parameters: Binding<VAlertParameters?>
     ) -> some View {
         vAlert(
             id: id,
-            uiModel: parameters.wrappedValue?.uiModel ?? .init(),
+            uiModel: uiModel,
             item: parameters,
             title: { $0.title },
             message: { $0.message },