+++ Sources/VComponents/Helpers/Architectural Pattern Helpers/Spinner/VSpinnerExtension_Parameters.swift
@@ -10,11 +10,12 @@ import VCore
 
 // MARK: - V Spinner Extension (Parameters)
 extension View {
-    /// Presents `VSpinner` when `VSpinnerParameters` is non-`nil`.
+    /// Presents `VContinuousSpinner` when `VSpinnerParameters` is non-`nil`.
     ///
     /// Done in the style of `View.progressView(parameters:)` from `VCore`.
     /// For additional info, refer to [documentation](https://github.com/VakhoKontridze/VCore/blob/main/Sources/VCore/Helpers/Architectural%20Pattern%20Helpers/SwiftUI/Progress%20View/ProgressViewExtension.swift) .
-    @ViewBuilder public func vSpinner(
+    @ViewBuilder public func vContinuousSpinner(
+        uiModel: VContinuousSpinnerUIModel = .init(),
         parameters: VSpinnerParameters?
     ) -> some View {
         switch parameters {
@@ -24,17 +25,27 @@ extension View {
         case let parameters?:
             self
                 .blocksHitTesting(parameters.isInteractionDisabled)
-                .overlay({
-                    Group(content: {
-                        switch parameters.spinnerType {
-                        case .continuous(let uiModel):
-                            VContinuousSpinner(uiModel: uiModel)
-                            
-                        case .dashed(let uiModel):
-                            VDashedSpinner(uiModel: uiModel)
-                        }
-                    })
-                }())
+                .overlay(VContinuousSpinner(uiModel: uiModel))
+        }
+    }
+    
+    /// Presents `VDashedSpinner` when `VSpinnerParameters` is non-`nil`.
+    ///
+    /// Done in the style of `View.progressView(parameters:)` from `VCore`.
+    /// For additional info, refer to [documentation](https://github.com/VakhoKontridze/VCore/blob/main/Sources/VCore/Helpers/Architectural%20Pattern%20Helpers/SwiftUI/Progress%20View/ProgressViewExtension.swift) .
+    @available(iOS 14.0, *)
+    @ViewBuilder public func vDashedSpinner(
+        uiModel: VDashedSpinnerUIModel = .init(),
+        parameters: VSpinnerParameters?
+    ) -> some View {
+        switch parameters {
+        case nil:
+            self
+            
+        case let parameters?:
+            self
+                .blocksHitTesting(parameters.isInteractionDisabled)
+                .overlay(VDashedSpinner(uiModel: uiModel))
         }
     }
 }