+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 General
 
-- Package now supports `iOS` `14.0`
+- Package now supports `iOS` `13.0`
 - `PresentationHost` is moved to [VCore](https://github.com/VakhoKontridze/VCore)
 - Demo app replaced with details component previews
 - Some colors are changed throughout the package
@@ -54,6 +54,11 @@ VAlert
 VContinuousSpinner
 
 - `borderWidth` is renamed to `thickness` in `VContinuousSpinnerUIModel`
+
+Helpers - Architectural Pattern Helpers
+
+- `VAlertUIModel` is moved from `VAlertParameters` to `View.vAlert(id:parameters)`
+- `SpinnerType` is removed from `VSpinnerParameters` and is moved to `View.vContinuousSpinner(parameters:)` and `View.vDashedSpinner(parameters:)`
 
 ### [3.2.0(29)](https://github.com/VakhoKontridze/VComponents/releases/tag/3.2.0) — *2023 03 09*